requirements.txt
# Uygulama kütüphaneleri

# BU DOSYA NE İŞE YARAR?
# Python projenizin hangi kütüphanelere bağımlı olduğunu belirtir.
# Dockerfile bunu kullanarak kurulumları yapacaktır.

# Eklenecek Kütüphaneler:
pymysql
pandas
matplotlib

python-dotenv
google-generativeai
fastapi
uvicorn[standard]
